requirements.txt
Flask==2.1.0
beautifulsoup4==4.10.0
selenium==3.141.0
soupsieve==2.3.1
urllib3==1.26.8
boto3==1.17.72
webdriver-manager
python-dotenv